Move StreamTask into a folder and split the dispatcher loop.
Colocate the run loops, emit path, checkpoint aligner, and WatermarkManager under runtime/stream_task; leave planner watermark config in runtime/watermark.
